데이터 서비스 클래스 추가, 컨트롤러 내 서비스 코드 분리
