+++ src/main/java/kr/ac/pknu/capstone/service/DataService.java
@@ -1,0 +1,27 @@
+package kr.ac.pknu.capstone.service;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import kr.ac.pknu.capstone.domain.Data.Data;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.stereotype.Service;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.List;
+
+
+@Service
+public class DataService {
+
+    @Autowired
+    ObjectMapper objectMapper;
+
+    public List<Data> readJsonData() throws IOException {
+        ClassPathResource resource = new ClassPathResource("data.json");
+        File file = resource.getFile();
+        return objectMapper.readValue(file, new TypeReference<List<Data>>() {});
+    }
+
+}
